pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webdominer"
version = "1.0.0"
description = "Open-web corpus generation from requirements specification documents."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Abdullaziz Hamad" }
]
dependencies = [
  "duckduckgo-search>=8.1.1",
  "trafilatura>=2.0.0",
  "playwright>=1.50.0",
  "keybert>=0.9.0",
  "sentence-transformers>=2.7.0",
  "scikit-learn>=1.3.0",
  "requests>=2.31.0",
  "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "black>=24.0.0",
  "isort>=5.13.0",
]

[project.scripts]
webdominer = "webdominer.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["webdominer*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"